+++ src/engine.ts
@@ -1,4 +1,4 @@
-import { AssetResources, EngineOptions, LoadedResources, Vector2 } from "./types";
+import { AssetResources, EngineOptions, LoadedResources, playAudioOptions, Vector2 } from "./types";
 import { EntityId, EntityManager, ComponentClass } from "entix-ecs";
 import { AssetLoader } from "./assetLoader";
 import { Transform, Sprite, EntityActive, Parent, Scene, Children, Rectangle, Shape, Circle, Triangle, Text, Alignment } from "./components";
@@ -6,6 +6,10 @@ import { renderingSystem } from "./systems/renderingSystem";
 import { Input } from "./input";
 import { buttonActionSystem } from "./systems/buttonActionSystem";
 import { alignmentSystem } from "./systems/alignmentSystem";
+import { EngineInternals } from "./internals/engineInternals";
+import { particleRemovalSystem } from "./systems/particleRemovalSystem";
+import { particleSimulationSystem } from "./systems/particleSimulationSystem";
+import { StorageManager } from "./storageManager";
 export class Engine {
     private canvas: HTMLCanvasElement | null = null;
     private ctx: CanvasRenderingContext2D | null = null;
@@ -24,13 +28,17 @@ export class Engine {
     private sceneEntities: EntityId[] = [];
     private currentSceneId: EntityId | null = null;
     private input: Input | null = null;
-    constructor(options: EngineOptions ={}) {
+    private domInteracted: boolean = false;
+    private onDOMInteractionFunctions: Function[] = [];
+    private storageManager: StorageManager | null = null;
+    constructor(options: EngineOptions = {}) {
         this.init(options);
         this.start().then(() => {
             requestAnimationFrame(this.update.bind(this));
         });
 
     };
+    //PRIVATES
     private init(options: EngineOptions) {// used to init all the important libs / files
         this.canvas = options.canvas ?? null;
         if (!this.canvas) {// create one
@@ -58,6 +66,8 @@ export class Engine {
         this.assetLoader = new AssetLoader(this.resources);
         this.input = new Input(this);
         if (!this.input) throw new Error("INPUT NOT FOUND!");
+        this.storageManager = new StorageManager();
+        if (!this.storageManager) throw new Error("STORAGE MANAGAER NOT FOUND!");
     };
     private async start(): Promise<void> {//other game stuff
         try {
@@ -77,6 +87,10 @@ export class Engine {
 
         this.ctx?.clearRect(0, 0, this.canvasWidth, this.canvasHeight);
 
+        this.waitForDOMInteraction();
+
+        particleRemovalSystem(this);
+        particleSimulationSystem(this);
         alignmentSystem(this);
         renderingSystem(this);
         buttonActionSystem(this);
@@ -88,6 +102,25 @@ export class Engine {
         this.input?.resetStates();
 
         requestAnimationFrame(this.update.bind(this));
+    };
+    private waitForDOMInteraction() {
+        const listener = () => {
+            if (this.domInteracted) return;
+            this.domInteracted = true;
+
+            // Trigger callbacks immediately in the same stack as the event
+            this.onDOMInteractionFunctions.forEach(fn => fn());
+            this.onDOMInteractionFunctions = []; // optional cleanup
+
+            console.log("DOM INTERACTION COMPLETE!");
+
+            // Remove listeners after interaction
+            window.removeEventListener("pointerdown", listener);
+            window.removeEventListener("keydown", listener);
+        };
+
+        window.addEventListener("pointerdown", listener);
+        window.addEventListener("keydown", listener);
     };
     //GETTER
     public getCanvas(): HTMLCanvasElement {
@@ -125,36 +158,42 @@ export class Engine {
             throw new Error("LOADED JSON RESOURCE " + key + " NOT FOUND!");
         return this.loadedResources.customJsonData[key] as T;
     };
-    public getSceneByName(name: string): EntityId | null {
+    public getSceneByName(name:string):EntityId | null;
+    public getSceneByName(name:string,strict:true):EntityId;
+    public getSceneByName(name:string,strict:false):EntityId | null;
+    public getSceneByName(name: string,strict?:boolean): EntityId | null {
+        const em:EntityManager = this.getEntityManager();
         for (const id of this.sceneEntities) {
-            const scene = this.entityManager?.getComponent(id, Scene);
-            if (scene && scene.name === name) return id;
-        }
+            const scene = em.getComponent(id, Scene,true);
+            if (scene.name === name) return id;
+        };
+        if(strict) throw new Error("SCENE ID NOT FOUND IN getSceneByName(). KEY: " + name);
         return null;
     };
     public getInput(): Input {
         if (!this.input) throw new Error("INPUT NOT FOUND!");
         return this.input;
     };
-    public getChildWithComponent<T>(id: EntityId, componentClass: ComponentClass<T>): EntityId | null {//Get first component with class
+    public getChildWithComponent<T>(id:EntityId,componentClass:ComponentClass<T>):EntityId | null;
+    public getChildWithComponent<T>(id:EntityId,componentClass:ComponentClass<T>,strict:true):EntityId;
+    public getChildWithComponent<T>(id:EntityId,componentClass:ComponentClass<T>,strict:false):EntityId | null;
+    public getChildWithComponent<T>(id: EntityId, componentClass: ComponentClass<T>,strict?:boolean): EntityId | null {//Get first component with class
         const em: EntityManager = this.getEntityManager();
-        const children = em.getComponent(id, Children);
-        if (!children)
-            throw new Error("CHILDREN COMPONENT NOT FOUND IN GETCHILDWITHCOMPONENT FN" + id);
+        const children = em.getComponent(id, Children,true);
         for (const childId of children.value) {
             const component = em.getComponent(childId, componentClass);
             if (component) {
                 return childId;
             }
         }
+        if(strict)
+            throw new Error("NO CHILD FOUND WITH THE REQUESTED COMPONENT IN getChildWithComponet(). ID: "+id+" COMPONENT:"+componentClass.name)
         return null;
     };
     public getChildrenWithComponent<T>(id: EntityId, componentClass: ComponentClass<T>): EntityId[] {//Returns all children with specified component
         const em: EntityManager = this.getEntityManager();
-        const children = em.getComponent(id, Children);
+        const children = em.getComponent(id, Children,true);
         let entityArr: EntityId[] = [];
-        if (!children)
-            throw new Error("CHILDREN COMPONENT NOT FOUND IN GETCHILDRENWITHCOMPONENT FN" + id);
         children.value.forEach((childId: EntityId) => {
             if (em.getComponent(childId, componentClass)) {
                 entityArr.push(childId);
@@ -162,14 +201,22 @@ export class Engine {
         });
         return entityArr;
     };
-    getRandomInt(min: number = 0, max: number = 1): number {
+    public getRandomInt(min: number = 0, max: number = 1): number {
         return Math.floor(Math.random() * (max - min + 1)) + min
     };
-    getRandomFloat(min: number = 0, max: number = 1): number {
+    public getRandomFloat(min: number = 0, max: number = 1): number {
         return Math.random() * (max - min) + min;
+    };
+    public getStorageManager(): StorageManager {
+        if (!this.storageManager) throw new Error("STORAGE MANAGAER NOT FOUND!");
+        return this.storageManager;
     };
     //FUNCTIONS
     public addUpdateFunction(fn: Function) {
+        if (this.updateFunctions.includes(fn)) {
+            console.warn("FAILED TO ADD UPDATE FN IN addUpdateFunction(), ALREADY CONTAINS FUNCTION !", fn);
+            return;
+        }
         this.updateFunctions.push(fn);
     };
     public removeUpdateFunction(fn: Function) {
@@ -179,6 +226,10 @@ export class Engine {
         }
     };
     public addStartFunction(fn: Function) {
+        if (this.startFunctions.includes(fn)) {
+            console.warn("FAILED TO ADD START FN IN addStartFunction(), ALREADY CONTAINS FUNCTION !", fn);
+            return;
+        }
         this.startFunctions.push(fn);
     };
     public removeStartFunction(fn: Function) {
@@ -188,12 +239,12 @@ export class Engine {
         }
     };
     public isEntityActive(id: EntityId): boolean {//checks whether the entity or it's parent chain is active or not
+        const em:EntityManager = this.getEntityManager();
         let currentEntity: EntityId | null = id;
 
         while (currentEntity !== null) {
-            const entityActiveComponent: EntityActive | null = this.entityManager?.getComponent(currentEntity, EntityActive) ?? null;
-            const parentComponent: Parent | null = this.entityManager?.getComponent(currentEntity, Parent) ?? null;
-            if (!entityActiveComponent) throw new Error("ENTITY DOES NOT HAVE ENTITYACTIVE COMPONENT IN ISENTITYACTIVE() " + id);
+            const entityActiveComponent  = em.getComponent(currentEntity, EntityActive,true);
+            const parentComponent: Parent | null = em.getComponent(currentEntity, Parent) ?? null;
             // Early return if entity is a Scene
             const sceneComponent = this.entityManager?.getComponent(currentEntity, Scene);
             if (sceneComponent) {
@@ -208,6 +259,9 @@ export class Engine {
         return true;
     };
     public addScene(id: EntityId) {//ads a scene,used upon creating a scene using scene template
+        const em:EntityManager = this.getEntityManager();
+        if(!em.hasComponent(id,Scene))
+             throw new Error("ENTITY DOES NOT HAS A SCENE COMPONENT IN addScene() YET YOU ARE TRYING TO ADD IT !")
         if (this.sceneEntities.includes(id)) {
             console.warn("THE SCENE YOU ARE TRYING TO ADD IS ALREADY ADDED! " + id);
             return;
@@ -215,13 +269,11 @@ export class Engine {
         this.sceneEntities.push(id);
     };
     public loadScene(id: EntityId) {//loads a scene
-        const sceneComponent = this.entityManager?.getComponent(id, Scene);
-        if (!sceneComponent) throw new Error("Scene component not found on entity: " + id);
+        const em:EntityManager = this.getEntityManager();
+        const sceneComponent = em.getComponent(id, Scene,true);
 
         for (const sceneEntity of this.sceneEntities) {// turn of all scenes
-            const entityActive = this.entityManager?.getComponent(sceneEntity, EntityActive);
-            if (!entityActive) throw new Error
-                ("SCENE DOES NOT HAVE ENTITY ACTIVE COMPONENT! WONT BE ABLE TO SWITCH TO THIS SCENE! " + sceneEntity);
+            const entityActive = em.getComponent(sceneEntity, EntityActive,true);
             entityActive.value = false;
         }
 
@@ -231,20 +283,16 @@ export class Engine {
         }
 
         this.currentSceneId = id;
-        const entityActive = this.entityManager?.getComponent(this.currentSceneId, EntityActive);
-        if (!entityActive) throw new Error("Target scene missing EntityActive component!");
+        const entityActive = em.getComponent(this.currentSceneId, EntityActive,true);
         entityActive.value = true;
         sceneComponent.onLoad();
     };
     public addParent(id: EntityId, parentId: EntityId) {// adds a parent to a entity
-        if (this.entityManager?.getComponent(id, Scene)) throw new Error// scenes dont have parent component!
+        const em:EntityManager = this.getEntityManager();
+        if (em.hasComponent(id, Scene)) throw new Error// scenes dont have parent component!
             ("CANT ASSIGN PARENT TO A SCENE! " + id);
-        const entityParentComponent = this.entityManager?.getComponent(id, Parent);
-        const parentChildrenComponent = this.entityManager?.getComponent(parentId, Children);
-        if (!entityParentComponent) throw new Error
-            ("ENTITY DOES NOT HAVE PARENT COMPONENT, AND YOU ARE TRYING TO ASSIGN A VALUE! " + id);
-        if (!parentChildrenComponent) throw new Error
-            ("ENTITY DOES NOT HAVE CHILDREN COMPONENT, AND YOU ARE TRYING TO ASSIGN A VALUE! " + id);
+        const entityParentComponent = em.getComponent(id, Parent,true);
+        const parentChildrenComponent = em.getComponent(parentId, Children,true);
         if (parentChildrenComponent.value.includes(id)) {
             console.warn
                 ("ENTITY ALREADY HAS THIS PARENT ASSIGNED! ARE YOU ASSIGNING THE SAME PARENT AGAIN ? " + id);
@@ -254,4 +302,36 @@ export class Engine {
         parentChildrenComponent.value.push(id);
 
     };
+    public playAudio(options: playAudioOptions) { //PLAYS AUDIO
+        if (!this.domInteracted) return;
+        const key: string = options.key;
+        const loop: boolean = options.loop ?? false;
+        const volume: number = options.volume ?? 1;
+        const audio: HTMLAudioElement = this.getAudio(options.key);
+        const cloneInstance: HTMLAudioElement = audio.cloneNode(false) as HTMLAudioElement;
+        cloneInstance.loop = loop;
+        cloneInstance.volume = volume;
+        cloneInstance.play();
+    };
+    public addOnDOMInteractionFunction(fn: Function) {
+        if (this.onDOMInteractionFunctions.includes(fn)) {
+            console.warn("FAILED TO ADD START FN IN addStartFunction(), ALREADY CONTAINS FUNCTION !", fn);
+            return;
+        };
+        this.onDOMInteractionFunctions.push(fn);
+    };
+    public removeEntityWithCleanup(id: EntityId) {
+        const engineInternals: EngineInternals = new EngineInternals(this);
+        engineInternals.unlinkFromParent(id);
+        engineInternals.removeChildrenReclusively(id);
+        this.entityManager?.removeEntity(id);
+    };
+    /* *EXPERIMENTAL* //NOT NEEDED
+    public mustGetComponent<T>(id:EntityId,comp:ComponentClass<T>):T{
+        const em:EntityManager = this.getEntityManager();
+        const c = em.getComponent(id,comp);
+        if(!c) throw new Error(comp + " COMPONENT NOT FOUND IN mustGetComponent()! ");
+        return c;
+    };
+    */
 };