+++ src/entityTemplates.ts
@@ -1,7 +1,7 @@
 import { EntityId, EntityManager } from "entix-ecs";
 import { Engine } from "./engine";
-import { Children, Circle, EntityActive, Parent, Rectangle, Scene, Shape, Sprite, Transform, Triangle, Text, Button, Alignment } from "./components";
-import { Vector2 } from "./types";
+import { Children, Circle, EntityActive, Parent, Rectangle, Scene, Shape, Sprite, Transform, Triangle, Text, Button, Alignment, ParticleContainer, Particle } from "./components";
+import { ParticleContainerOptions, Vector2 } from "./types";
 
 export class EntityTemplates {
     private engine: Engine;
@@ -93,10 +93,10 @@ export class EntityTemplates {
             }),
             blocksInput: false
         }));
-        const id2:EntityId = this.createTextEntity(textContent ?? 'Text',id);
-        this.em.addComponent(id2,Alignment, new Alignment({
-            alignmentHorizontal:'center',
-            alignmentVertical:'middle'
+        const id2: EntityId = this.createTextEntity(textContent ?? 'Text', id);
+        this.em.addComponent(id2, Alignment, new Alignment({
+            alignmentHorizontal: 'center',
+            alignmentVertical: 'middle'
         }));
         return id;
     };
@@ -112,5 +112,55 @@ export class EntityTemplates {
             blocksInput: false
         }));
         return id;
+    };
+    createParticleContainerEntity(options: ParticleContainerOptions): { id: EntityId, instance: Function } {
+        const id: EntityId = this.createEmptyEntity();
+        this.em.addComponent(id, ParticleContainer, new ParticleContainer(options));
+
+        const run: Function = () => {
+            const transform = this.em.getComponent(id, Transform,true);
+            const particleContainer = this.em.getComponent(id, ParticleContainer,true);
+
+            const pos: Vector2 = transform.globalPosition.position;
+            const img: HTMLImageElement = particleContainer.img;
+
+            //ATTACH THE VALUES
+            for (let i = 0; i < particleContainer.maxNumber; i++) { //CREATE CHILD PARTICLES
+                const scale: Vector2 = {
+                    x: this.engine.getRandomFloat(particleContainer.minScaleRange.x, particleContainer.maxScaleRange.x),
+                    y: this.engine.getRandomFloat(particleContainer.minScaleRange.y, particleContainer.maxScaleRange.y)
+                };
+                const speed: number = this.engine.getRandomFloat(particleContainer.minSpeedRange, particleContainer.maxSpeedRange);
+                const lifeTime: number = this.engine.getRandomFloat(particleContainer.minLifeTimeRange, particleContainer.maxLifeTimeRange);
+                const alpha: number = this.engine.getRandomFloat(particleContainer.minAlphaRange, particleContainer.maxAlphaRange);
+                const rotation: number = this.engine.getRandomFloat(particleContainer.minRotationRange, particleContainer.maxRotationRange);
+                const angle: number = Math.random() * Math.PI * 2;
+                const velocity: Vector2 = {
+                    x: Math.cos(angle) * speed,
+                    y: Math.sin(angle) * speed
+                };
+
+                const childId: EntityId = this.createSpriteEntity(img, img.width, img.height, id);
+                this.em.addComponent(childId, Particle, new Particle({
+                    lifeTime: lifeTime,
+                    velocity: velocity,
+                    alphaReductionRate: particleContainer.alphaReductionRate
+                }));
+
+                const childTransform = this.em.getComponent(childId, Transform,true);
+                const childSprite = this.em.getComponent(childId, Sprite,true);
+              
+                childTransform.scale.value = scale;
+                childTransform.globalPosition.position = {
+                    x: pos.x,
+                    y: pos.y
+                };
+                
+                childSprite.alpha = alpha;
+                childSprite.rotation = rotation;
+                
+            }
+        };
+        return { id: id, instance: run };
     }
 };